feat/configure: Refactor write func, delay write func call after confirm
